created pages directory
